requirements-container.txt
Flask==2.3.3
azure-ai-projects==1.0.0
azure-identity==1.15.0
azure-ai-agents==1.0.0
gunicorn==21.2.0
requests==2.31.0
Werkzeug==2.3.7